fix(post): Save cache on cache miss
getState always returns a string, so the boolean values we pass to
saveState are coerced to "false" and "true." Both of these are truthy
values, so the action always incorrectly stated that a cache hit
occurred without ever saving the cache. Test explicitly for the string
"true" rather than for truthiness so that the cache is saved on miss.

When saving Docker images on cache miss, redirect xargs --verbose output
from stderr to stdout since it typically isn't an error.