chore(Dependabot): Add Yarn
Automatically update Yarn dependencies via Dependabot. Dependabot
recently enabled Yarn v2+ support in production behind a feature flag.
We have the privilege of being part of the first batch of external repos
to receive access! Use the fix prefix in commit messages for bumps to
production dependencies so Commitizen bumps the patch version. Use the
chore prefix in commit messages for bumps to development dependencies so
that Commitizen doesn't bump the project version.