docs(README): Clarify how to create a cache key
Replace hashFiles(...) with hashFiles(paths) to make it clear that ...
isn't a special symbol and that the user ought to pass one or more
filepaths to hashFiles. Explicitly document that the user must select an
appropriate key and cannot copy the example verbatim. Link to the
specific section of the official GitHub cache action that documents how
to create a cache key.