ci(CSpell): Add win32 dict as dev dependency
Eliminate the need to manually install the CSpell win32 dict locally in
order for the VS Code CSpell extension to have access to that
dictionary. Keep the installation pre-command for now; it is a
workaround for a bug in MegaLinter where CSpell fails to import the
win32 dict when it is run via MegaLinter.