feat(action): Verify no matching cache before save
Prior to saving cache, verify that there is no cache with a matching key
already cached. Previously, we checked if a cache key matching our cache
was present in the load step; if not, we would then proceed to attempt
to save the cache in the save step (assuming we were not in read-only
mode and that there were indeed new Docker images to save). This was
problematic when the action was run in parallel; in that case, there
were multiple, unnecessary cache save attempts when the cache initially
missed but was subsequently saved.