feat(docker): Support dangling Docker images
Docker images that were overwritten by a new one
with the same name and tag do not have a name or
tag. Docker images that were pulled by digest do
not have a tag. Support the former by falling back
on their ID and the latter by using their name.

These changes prevent the action from crashing on
such images. Continue using both the name and tag
for images that have them so that they are
retained upon restoring the cache.